liufeng 2019-04-25 20：:53 提交

+++ src/main/java/com/woniu/yoga/user/pojo/User.java
@@ -43,9 +43,9 @@ public class User {
 
     private String userHeadimg;
 
-    private Float longitude;
+    private double longitude;
 
-    private Float latitude;
+    private double latitude;
 
     private String userLocation;
